requirements.txt
Django==2.2.17
mysqlclient==2.0.3
Pillow==2.2.1

